systemd: override batman
batman MUST start earlier on this device, by default on boot half the cores are offline and frequency is set to the lowest possible value (why? still up to discussion).

batman must start early to reset everything back to maximum performance. another solution would be to have a script set all of these on boot and leave batman as is, but batman already has whatever is needed so doesn't hurt so just use what we have

default target of batman is graphical, but graphical can take a while to actually completely finish leaving the device in a very unusable state for the first 10 seconds of phosh starting

we can start batman after multi user and after phosh then add a small sleep to make sure it starts early but not too early.